pyproject.toml
[tool.poetry]
name = "twitter_sentiment"
version = "0.1.1"
description = "Project to analyze tweets sentiment"
authors = ["arctickey <ffilipchrzuszcz1@gmail.com>"]
repository = "https://github.com/fpgmaas/twitter-sentiment"
documentation = "https://fpgmaas.github.io/twitter-sentiment/"
readme = "README.md"
packages = [
  {include = "twitter_sentiment"}
]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
tweepy = "^4.10.0"
pre-commit = "^2.20.0"
kafka-python = "^2.0.2"
apache-airflow = "^2.3.3"
psycopg2-binary = "^2.9.3"
pyspark = "^3.3.0"
pandas = "^1.4.3"
transformers = "^4.21.2"
torch = "^1.12.1"
sklearn = "^0.0"


[tool.poetry.dev-dependencies]
black = "^22.6.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
pytest = "^7.1.2"
mkdocs = "^1.3.0"
mkdocs-material = "^8.3.8"
mkdocstrings = "^0.18.1"
mypy = "^0.961"
ipykernel = "^6.15.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
include = '\.pyi?$'
target-version = ['py39']
fast = true
exclude = '''
(
  /(                        # exclude a few common directories in the
    \.git                   # root of the project
    | \.pytest_cache
    | python-venv
    | \.venv
    | build
    | dist
    | \.tox
  ))
'''

[tool.isort]
profile = "black"

[tool.mypy]
ignore_missing_imports = "True"
disallow_untyped_defs = "True"
disallow_any_unimported = "True"
no_implicit_optional = "True"
check_untyped_defs = "True"
warn_return_any = "True"
warn_unused_ignores = "True"
show_error_codes = "True"
exclude = [
    '\.venv',
    'tests'
]
